fix compilation error with C23 compiler

+++ src/engine/net_server.c
@@ -827,7 +827,7 @@ static void NET_SV_ParseGameData(net_packet_t* packet, net_client_t* client)
 
 	// Read header
 
-	if (!NET_ReadInt8(packet, (int8_t *)&ackseq)
+	if (!NET_ReadInt8(packet, (int *)&ackseq)
 		|| !NET_ReadInt8(packet, &seq)
 		|| !NET_ReadInt8(packet, &num_tics))
 	{
@@ -946,7 +946,7 @@ static void NET_SV_ParseGameDataACK(net_packet_t* packet, net_client_t* client)
 
 	// Read header
 
-	if (!NET_ReadInt8(packet, (int8_t *)&ackseq))
+	if (!NET_ReadInt8(packet, (int *)&ackseq))
 	{
 		return;
 	}
